need to use BigDecimal in java to map correctly to decimal values in mysql. updating service and related files

+++ intellij_project/src/main/java/it/polimi/telco_webapp/entities/Service.java
@@ -5,6 +5,7 @@ package it.polimi.telco_webapp.entities;
 import it.polimi.telco_webapp.auxiliary.ServiceType;
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Table(name = "service", schema = "telco_db")
@@ -21,14 +22,14 @@ public class Service {
     @Enumerated(EnumType.STRING)
     private ServiceType type;
 
-    @Column(name = "base_price1", nullable = false, precision = 2)
-    private Double basePrice1;
+    @Column(name = "base_price1", nullable = false)
+    private BigDecimal basePrice1;
 
-    @Column(name = "base_price2", nullable = false, precision = 2)
-    private Double basePrice2;
+    @Column(name = "base_price2", nullable = false)
+    private BigDecimal basePrice2;
 
-    @Column(name = "base_price3", nullable = false, precision = 2)
-    private Double basePrice3;
+    @Column(name = "base_price3", nullable = false)
+    private BigDecimal basePrice3;
 
     @Column(name = "gig_included")
     private Integer gigIncluded;
@@ -39,14 +40,14 @@ public class Service {
     @Column(name = "sms_included")
     private Integer smsIncluded;
 
-    @Column(name = "gig_extra", precision = 2)
-    private Double gigExtra;
+    @Column(name = "gig_extra")
+    private BigDecimal gigExtra;
 
-    @Column(name = "min_extra", precision = 2)
-    private Double minExtra;
+    @Column(name = "min_extra")
+    private BigDecimal minExtra;
 
-    @Column(name = "sms_extra", precision = 2)
-    private Double smsExtra;
+    @Column(name = "sms_extra")
+    private BigDecimal smsExtra;
 
     @ManyToMany
     @JoinTable(name = "service_bundles", joinColumns = @JoinColumn(name = "service_id"),
@@ -55,12 +56,12 @@ public class Service {
 
 
 
-    public Double getSmsExtra() {return smsExtra;}
-    public Double getMinExtra() {return minExtra;}
-    public Double getGigExtra() {return gigExtra;}
-    public void setSmsExtra(Double smsExtra) {this.smsExtra = smsExtra;}
-    public void setMinExtra(Double minExtra) {this.minExtra = minExtra;}
-    public void setGigExtra(Double gigExtra) {this.gigExtra = gigExtra;}
+    public BigDecimal getSmsExtra() {return smsExtra;}
+    public BigDecimal getMinExtra() {return minExtra;}
+    public BigDecimal getGigExtra() {return gigExtra;}
+    public void setSmsExtra(BigDecimal smsExtra) {this.smsExtra = smsExtra;}
+    public void setMinExtra(BigDecimal minExtra) {this.minExtra = minExtra;}
+    public void setGigExtra(BigDecimal gigExtra) {this.gigExtra = gigExtra;}
 
     public Integer getSmsIncluded() {return smsIncluded;}
     public Integer getMinIncluded() {return minIncluded;}
@@ -69,12 +70,12 @@ public class Service {
     public void setMinIncluded(Integer minIncluded) {this.minIncluded = minIncluded;}
     public void setGigIncluded(Integer gigIncluded) {this.gigIncluded = gigIncluded;}
 
-    public Double getBasePrice1() {return basePrice1;}
-    public Double getBasePrice2() {return basePrice2;}
-    public Double getBasePrice3() {return basePrice3;}
-    public void setBasePrice1(Double basePrice1) {this.basePrice1 = basePrice1;}
-    public void setBasePrice2(Double basePrice2) {this.basePrice2 = basePrice2;}
-    public void setBasePrice3(Double basePrice3) {this.basePrice3 = basePrice3;}
+    public BigDecimal getBasePrice1() {return basePrice1;}
+    public BigDecimal getBasePrice2() {return basePrice2;}
+    public BigDecimal getBasePrice3() {return basePrice3;}
+    public void setBasePrice1(BigDecimal basePrice1) {this.basePrice1 = basePrice1;}
+    public void setBasePrice2(BigDecimal basePrice2) {this.basePrice2 = basePrice2;}
+    public void setBasePrice3(BigDecimal basePrice3) {this.basePrice3 = basePrice3;}
 
 
     public ServiceType getServiceType() {return type; }
